[fixup] remove subschema_module, introduce Schema:new_subschema
With this fixup commit, requiring subschema modules is no longer the concern
of the schema library. Instead, it includes a new_subschema method to a
Schema, which takes a name and a schema definition table.

The actual (pre-)loading of plugin subschemas will be performed at a different
layer than the schema library.